# Refactor Frame scans